Feat: 리포트가 비어있을때 화면 ui 완성 & 리포트 리스트 코드 정리
