+++ Spender/app/src/main/java/com/example/spender/feature/report/ui/component/EmptyReport.kt
@@ -1,0 +1,52 @@
+package com.example.spender.feature.report.ui.component
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import com.example.spender.R
+import com.example.spender.ui.theme.LightFontColor
+import com.example.spender.ui.theme.Typography
+
+@Composable
+fun EmptyReport(paddingValues: PaddingValues){
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(paddingValues)
+            .padding(32.dp),
+        contentAlignment = Alignment.Center
+    ) {
+        Column(horizontalAlignment = Alignment.CenterHorizontally) {
+            Image(
+                painter = painterResource(id = R.drawable.spender_sad),
+                contentDescription = "지출이",
+                modifier = Modifier.size(256.dp)
+            )
+
+            Text(
+                text = "등록된 리포트가 없어요.",
+                style = Typography.bodyLarge,
+                color = LightFontColor
+            )
+            Spacer(modifier = Modifier.height(16.dp))
+            Text(
+                text = "지출을 등록하면 자동으로 리포트가 생성돼요!",
+                style = Typography.bodyMedium,
+                color = LightFontColor
+            )
+            Spacer(modifier = Modifier.height(40.dp))
+        }
+    }
+}
