+++ Spender/app/src/main/java/com/example/spender/feature/report/ui/component/FeedbackSection.kt
@@ -36,7 +36,7 @@ fun FeedbackSection(){
         // TODO : 말풍선 형식으로 (예산 보다 아꼈으면 웃는 지출이 / 더 쓰면 화난 지출이)
         FeedbackBox(
             text = "이번 달은 외식비가 많았어요.\n다음 달에는 줄여볼까요?",
-            imageResId = R.drawable.spender
+            imageResId = R.drawable.spender_happy
         )
 
         Spacer(modifier = Modifier.height(24.dp))
