+++ Spender/app/src/main/java/com/example/spender/feature/report/ui/component/ReportTopAppBar.kt
@@ -1,0 +1,62 @@
+package com.example.spender.feature.report.ui.component
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.KeyboardArrowLeft
+import androidx.compose.material.icons.automirrored.filled.KeyboardArrowRight
+import androidx.compose.material.icons.filled.DateRange
+import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.alpha
+import androidx.compose.ui.unit.dp
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ReportTopAppBar(year: Int, onPrev: () -> Unit, onNext: () -> Unit) {
+    CenterAlignedTopAppBar(
+        title = {
+            Row(
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                IconButton(onClick = onPrev, modifier = Modifier.size(36.dp)) {
+                    Icon(
+                        Icons.AutoMirrored.Filled.KeyboardArrowLeft,
+                        contentDescription = "이전 년도"
+                    )
+                }
+
+                Text(
+                    text = "${year}년",
+                    modifier = Modifier.padding(horizontal = 8.dp)
+                )
+
+                IconButton(
+                    onClick = onNext, // TODO : 리포트가 있는 경우에만 이동가능하도록 수정
+                    modifier = Modifier
+                        .size(36.dp)
+                        .alpha(if (year < 2025) 1f else 0f),
+                    enabled = true
+                ) {
+                    Icon(
+                        Icons.AutoMirrored.Filled.KeyboardArrowRight,
+                        contentDescription = "다음 년도"
+                    )
+                }
+            }
+        },
+        actions = {
+//            IconButton(onClick = { // TODO: 캘린더
+//            }) {
+//                Icon(Icons.Default.DateRange, contentDescription = "캘린더")
+//            }
+        }
+    )
+}
